ms_auth/pyproject.toml
[project]
name = "ms-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-azure-auth>=5.2.0",
    "msal>=1.35.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
    "python-jose[cryptography]>=3.3.0",
]
